+++ app/login/action.ts
@@ -1,0 +1,22 @@
+import { createClient } from "@/utils/supabase/server";
+import { redirect } from "next/navigation";
+
+export const signIn = async (formData: FormData) => {
+  "use server";
+
+  const email = formData.get("email") as string;
+  const password = formData.get("password") as string;
+  const supabase = createClient();
+
+  const { error } = await supabase.auth.signInWithPassword({
+    email,
+    password,
+  });
+
+  if (error) {
+    console.log(error);
+    return redirect("/login?message=Could not authenticate user");
+  }
+
+  return redirect("/protected");
+};
