+++ app/login/page.tsx
@@ -2,33 +2,14 @@ import Link from "next/link";
 import { headers } from "next/headers";
 import { createClient } from "@/utils/supabase/server";
 import { redirect } from "next/navigation";
-import { SubmitButton } from "./submit-button";
 import LoginForm from "@/components/LoginForm";
+import { signIn } from "./action";
 
 export default function Login({
   searchParams,
 }: {
   searchParams: { message: string };
 }) {
-  const signIn = async (formData: FormData) => {
-    "use server";
-
-    const email = formData.get("email") as string;
-    const password = formData.get("password") as string;
-    const supabase = createClient();
-
-    const { error } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    });
-
-    if (error) {
-      return redirect("/login?message=Could not authenticate user");
-    }
-
-    return redirect("/protected");
-  };
-
   const signUp = async (formData: FormData) => {
     "use server";
 
@@ -52,5 +33,5 @@ export default function Login({
     return redirect("/login?message=Check email to continue sign in process");
   };
 
-  return <LoginForm />;
+  return <LoginForm login={signIn} />;
 }
