+++ app/protected/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function page() {
+  return <div>page Protected</div>;
+}
+
+export default page;
