+++ utils/supabase/middleware.ts
@@ -57,12 +57,17 @@ export const updateSession = async (request: NextRequest) => {
             });
           },
         },
-      },
+      }
     );
 
     // This will refresh session if expired - required for Server Components
     // https://supabase.com/docs/guides/auth/server-side/nextjs
-    await supabase.auth.getUser();
+    const user = await supabase.auth.getUser();
+
+    //redirect if not logged in
+    if (request.nextUrl.pathname.startsWith("/protected") && user.error) {
+      return NextResponse.redirect(new URL("/login", request.url));
+    }
 
     return response;
   } catch (e) {
